+++ src/Http/Controllers/LlmsTxtController.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Laradocs\Http\Controllers;
+
+use Illuminate\Http\Response;
+use Laradocs\Laradocs;
+
+final class LlmsTxtController
+{
+    public function __construct(
+        private readonly Laradocs $laradocs,
+    ) {}
+
+    public function __invoke(): Response
+    {
+        return new Response($this->laradocs->llmsTxt(), 200, [
+            'Content-Type' => 'text/plain; charset=UTF-8',
+        ]);
+    }
+}